core: split core/arch/arm/include/mm/core_mmu.h
Splits core/arch/arm/include/mm/core_mmu.h into one generic and one
architecture specific file.

Acked-by: Marouene Boubakri <marouene.boubakri@nxp.com>
